[core] don't release mainaction before stopping workers
This reverts commit f2ade38923221bb74321c16d9e6254a4e27428a6.

Running workers will try to handle connections, and thus changing
mainaction without locking is a race condition, and also we don't handle
a NULL mainaction properly and segfault.

Given running connections will keep mainaction alive anyway, it is
unclear what the purpose of releasing it early was in the first place.

Change-Id: I2266958520c3cfca28c5c19574abdf44ddf6b0bc